variable type fix according to w32 api from spintz.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@37 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CFileList.cpp
@@ -40,7 +40,7 @@ CFileList::CFileList()
 	Path = tmp;
 
 	struct _finddata_t c_file;
-	s32 hFile;
+	intptr_t hFile;
 	FileEntry entry;
 
 	if( (hFile = _findfirst( "*", &c_file )) != -1L )
